requirements.txt
fastapi
sqlmodel
sqlalchemy-utils
aiosqlite
strawberry-graphql[fastapi]
